server/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
PyGithub==2.5.0
requests>=2.31.0
httpx>=0.27.2